feat: Added support for execution on remotes 🌎
Executables (compiled script templates) can now be executed locally, over ssh, or on a docker container. For this, the concept of a 'process' (running or completed executable) has been drastically changed. It is now an interface, with Local SSH and Docker implementing it. In the process, the code for output parsing was seperated.